Version 3.3.0
Support for the Joomla update system.
Better handling for when a user closes the shadowbox while an image is still loading.
PHP 7 support - There weren't any errors, but have now removed a deprecated API call.
Support for a new addon! There's now a social addon, which lets you add buttons to the shadowbox to share directly to Facebook, Twitter, Google+ or Tumblr.
Support for anchor links. Anchor links allow directly linking to an image in the gallery. When the page loads, it scrolls to that image in the gallery, and opens it in the shadowbox seamlessly.
Update to the image sizing algorithm, to keep images much closer to the target row height, for a cleaner, more organized look.
Now supports touch gestures! Mobile users can now swipe left or right to navigate through images in the shadowbox (note that if you're updating from an older version, you may need to change to the newer shadowbox in the plugin's settings to get this feature).
Some code housekeeping and minor bug fixes